feat(#2190): replace direct brick imports in services/ with Protocol boundaries
Move shared types to tier-neutral contracts/ layer and introduce narrow
Protocols for Memory brick dependencies, enforcing LEGO Architecture §3.3
("bricks don't know about each other").

- Create contracts/llm_types.py, contracts/rebac_types.py,
  contracts/search_types.py with zero nexus.* runtime imports
- Add MemoryPermissionCheckerProtocol and EntityResolverProtocol to
  services/protocols/memory.py (Issue #2190)
- Update 18 import sites across services/ to use contracts/ or Protocols
- Convert 5 brick modules to backward-compat re-export shims
- Add 15 import identity tests, 11 Protocol isolation tests with fakes,
  2 Protocol compliance tests

Performance: import time improved 28% (191ms → 137ms), zero regression
in permission benchmarks.
