+++ src/nexus/rebac/cross_zone.py
@@ -1,18 +1,13 @@
 """Cross-zone sharing constants for ReBAC federation.
 
+Backward-compat shim (Issue #2190): Canonical location is
+``nexus.contracts.rebac_types``. This module re-exports for existing importers.
+
 These constants define which relations are allowed to span zone boundaries.
 Cross-zone sharing is a federation-specific policy concept
 (KERNEL-ARCHITECTURE §3, federation-memo §6).
 """
 
-from __future__ import annotations
-
-# Relations that are allowed to cross zone boundaries.
-# These relations can link subjects and objects from different zones.
-CROSS_ZONE_ALLOWED_RELATIONS: frozenset[str] = frozenset(
-    {
-        "shared-viewer",  # Read access via cross-zone share
-        "shared-editor",  # Read + Write access via cross-zone share
-        "shared-owner",  # Full access via cross-zone share
-    }
+from nexus.contracts.rebac_types import (  # noqa: F401
+    CROSS_ZONE_ALLOWED_RELATIONS as CROSS_ZONE_ALLOWED_RELATIONS,
 )
