+++ src/nexus/services/permissions/graph/traversal.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING, Any
 
 from sqlalchemy import or_, select
 
-from nexus.rebac.domain import WILDCARD_SUBJECT, Entity
+from nexus.contracts.rebac_types import WILDCARD_SUBJECT, Entity
 from nexus.storage.models.permissions import ReBACTupleModel as RT
 
 if TYPE_CHECKING:
