+++ src/nexus/services/permissions/graph/zone_traversal.py
@@ -23,8 +23,9 @@ from typing import TYPE_CHECKING, Any
 
 from sqlalchemy import or_, select
 
-from nexus.rebac.domain import WILDCARD_SUBJECT, Entity
-from nexus.rebac.types import (
+from nexus.contracts.rebac_types import (
+    WILDCARD_SUBJECT,
+    Entity,
     GraphLimitExceeded,
     GraphLimits,
     TraversalStats,
