+++ src/nexus/services/permissions/tuples/repository.py
@@ -23,7 +23,7 @@ from typing import TYPE_CHECKING, Any
 from sqlalchemy import or_, select
 
 from nexus.constants import ROOT_ZONE_ID
-from nexus.rebac.domain import WILDCARD_SUBJECT, Entity
+from nexus.contracts.rebac_types import WILDCARD_SUBJECT, Entity
 from nexus.storage.models.permissions import (
     ReBACTupleModel as RT,
 )
