+++ src/nexus/services/search_grep_mixin.py
@@ -19,12 +19,12 @@ import os
 import re
 from typing import TYPE_CHECKING, Any, cast
 
-from nexus.core import glob_fast, grep_fast, trigram_fast
-from nexus.core.rpc_decorator import rpc_expose
-from nexus.search.strategies import (
+from nexus.contracts.search_types import (
     GREP_PARALLEL_WORKERS,
     SearchStrategy,
 )
+from nexus.core import glob_fast, grep_fast, trigram_fast
+from nexus.core.rpc_decorator import rpc_expose
 
 logger = logging.getLogger(__name__)
